Add reviews

+++ src/main/java/julia/books/domain/books/SearchResult.java
@@ -9,5 +9,5 @@ public class SearchResult<T> {
     List<T> content;
     int page;
     int size;
-    int totalElements;
+    long totalElements;
 }
